iotpro/iotproject11/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
PyJWT==2.10.1
